Feat/add extension

+++ app/api/content/[id]/route.ts
@@ -2,18 +2,21 @@ import { NextRequest, NextResponse } from "next/server";
 import { authOptions } from "@/libs/auth";
 import { getServerSession } from "next-auth/next";
 import { prisma } from "@/db";
+import ratelimit from "@/libs/ratelimit";
 export async function DELETE(
   req: NextRequest,
   { params }: { params: Promise<{ id: string }> }
 ) {
-  const id = (await params).id
-
+  const id = (await params).id;
+  const ip = req.headers.get("x-forwarded-for") || "127.0.0.1";
   const session = await getServerSession(authOptions);
 
   if (!session) {
     return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
   }
-
+  if (!ratelimit(ip)) {
+    return NextResponse.json({ message: "Too many request!" }, { status: 429 });
+  }
   try {
     const deletedContent = await prisma.content.delete({
       where: {
